+++ resources/views/components/create.blade.php
@@ -1,5 +1,0 @@
-@props(['redirect'])
-
-<div>
-    <livewire:filament-multifactor-passkeys-register :redirect-url="$redirect" />
-</div>